Separate canvas framework from user canvas files
Move templates, base classes, views, package manager, and canvas_manager
into src/canvas_runtime/ so canvas/ only contains user-authored *.py
modules. Docker volume mounts of canvas/ no longer overlay framework
code, which simplifies persisting canvases across image updates.
