+++ src/canvas_runtime/views/_base_view.py
