+++ src/canvas_runtime/views/_new_view_template.py
