+++ src/canvas_runtime/views/analog_clock.py
