+++ src/canvas_runtime/views/circle.py
