+++ src/canvas_runtime/views/image_view.py
