+++ src/canvas_runtime/views/square.py
