+++ src/canvas_runtime/views/triangle.py
